python/reranker/requirements.txt
flask==2.3.3
pyyaml==6.0.1
torch==2.0.1
transformers==4.32.1 